+++ app/src/main/java/com/dobrihlopez/financeassistant/feature/accounts/data/repo/AccountsRepositoryImpl.kt
@@ -1,6 +1,8 @@
-package com.dobrihlopez.financeassistant.feature.accounts.data
+package com.dobrihlopez.financeassistant.feature.accounts.data.repo
 
-import com.dobrihlopez.financeassistant.feature.accounts.data.model.UpdateAccountRequest
+import com.dobrihlopez.financeassistant.feature.accounts.data.network.AccountsApiService
+import com.dobrihlopez.financeassistant.feature.accounts.domain.AccountsRepository
+import com.dobrihlopez.financeassistant.feature.accounts.data.network.UpdateAccountRequest
 import com.dobrihlopez.financeassistant.feature.accounts.domain.UserAccountDetailed
 import com.dobrihlopez.financeassistant.feature.accounts.domain.toDomain
 import javax.inject.Inject
@@ -13,4 +15,4 @@ class AccountsRepositoryImpl @Inject constructor(
 
     override suspend fun updateAccount(id: Int, name: String, balance: String, currency: String): UserAccountDetailed =
         api.updateAccount(id, UpdateAccountRequest(name, balance, currency)).toDomain()
-} 
+}
