+++ app/src/main/java/com/dobrihlopez/financeassistant/feature/accounts/domain/AccountsRepository.kt
@@ -1,8 +1,6 @@
-package com.dobrihlopez.financeassistant.feature.accounts.data
-
-import com.dobrihlopez.financeassistant.feature.accounts.domain.UserAccountDetailed
+package com.dobrihlopez.financeassistant.feature.accounts.domain
 
 interface AccountsRepository {
     suspend fun getAccounts(): List<UserAccountDetailed>
     suspend fun updateAccount(id: Int, name: String, balance: String, currency: String): UserAccountDetailed
-} 
+}
