+++ app/src/main/java/com/dobrihlopez/financeassistant/feature/accounts/presentation/AccountsStore.kt
@@ -10,7 +10,7 @@ import com.dobrihlopez.financeassistant.feature.accounts.domain.UserAccountDetai
 import kotlinx.serialization.Serializable
 import javax.inject.Inject
 import kotlinx.coroutines.launch
-import com.dobrihlopez.financeassistant.feature.accounts.domain.GetFirstAccountUseCase
+import com.dobrihlopez.financeassistant.feature.accounts.domain.usecase.GetFirstAccountUseCase
 import com.dobrihlopez.financeassistant.feature.accounts.domain.UpdateAccountUseCase
 
 interface AccountsStore: Store<AccountsStore.Intent, AccountsStore.AccountScreenState, Nothing> {
