+++ app/src/main/java/com/dobrihlopez/financeassistant/feature/transaction/expenses/domain/ExpenseRepository.kt
@@ -1,4 +1,4 @@
-package com.dobrihlopez.financeassistant.core.domain.expenses
+package com.dobrihlopez.financeassistant.feature.transaction.expenses.domain
 
 interface ExpenseRepository {
     suspend fun getExpenseList(): List<Expense>
