--- app/src/main/java/com/dobrihlopez/financeassistant/feature/transaction/history/di/HistoryModule.kt
@@ -1,2 +1,0 @@
-package com.dobrihlopez.financeassistant.feature.transaction.history.di
-
